requirements.txt
flask
flask-sqlalchemy
flask-login
authlib
requests
python-dotenv